pertama

+++ database/migrations/2025_01_06_013129_create_produks_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('gambar');
             $table->string('deskripsi');
             $table->integer('stok');
-            $table->integer('harga');
+            $table->decimal('harga');
             $table->timestamps();
         });
     }
